+++ src/lib/components/Page.svelte
@@ -1,0 +1,20 @@
+<script lang="ts">
+	import type { Snippet } from 'svelte';
+
+	let { title, children }: { title: string; children: Snippet } = $props();
+</script>
+
+<svelte:window {title} />
+
+<div class="page">
+	{@render children()}
+</div>
+
+<style>
+	.page {
+		display: flex;
+		flex-direction: column;
+		gap: var(--spacing);
+		padding: var(--spacing);
+	}
+</style>