+++ src/routes/scholar/[id]/Name.svelte
@@ -1,5 +1,4 @@
 <script lang="ts">
-	import Note from '$lib/components/Note.svelte';
 	import EditableText from '$lib/components/EditableText.svelte';
 	import { getDB } from '$lib/data/Database';
 	import type Scholar from '$lib/data/Scholar.svelte';
@@ -15,15 +14,14 @@
 	}
 </script>
 
-<h1>
+<h2>
 	{#if editable}<EditableText
 			text={scholar.getName() ?? ''}
-			label="name"
 			placeholder="name"
 			change="Change name"
 			save="Save name"
 			empty={anonymous}
 			edit={update}
 		/>{:else}{scholar.getName() ?? anonymous}{/if}
-</h1>
+</h2>
 Joined {new Date(scholar.getJoined()).toLocaleDateString()}