+++ CHANGELOG.md
@@ -3,6 +3,12 @@
 Hi! This is where we document all notable changes, including bug fixes, enhancements, and dependency updates.
 Dates should be in`YYYY-MM-DD` format and versions are in [semantic versioning](http://semver.org/) format.
 
+## 0.0.5 2024-10-06
+
+### Added
+
+- Currency name and description display and editing.
+
 ## 0.0.4 2024-09-28
 
 ### Added